Wip devoluciones #1qeqwe7

+++ backend/src/main/java/org/foodmonks/backend/Restaurante/Restaurante.java
@@ -25,7 +25,7 @@ public class Restaurante extends Usuario {
 
     private Float calificacion;
     private String nombreRestaurante;
-    private Integer rut;
+    private Long rut;
     @OneToOne(cascade=CascadeType.ALL)
     private Direccion direccion;
     @Enumerated(value = EnumType.STRING)
@@ -48,7 +48,7 @@ public class Restaurante extends Usuario {
         super();
     }
 
-    public Restaurante(String nombre, String apellido, String correo, String contrasenia, LocalDate fechaRegistro, Float calificacion, String nombreRestaurante, Integer rut, Direccion direccion, EstadoRestaurante estado, Integer telefono, String descripcion, String cuentaPaypal, String imagen) {
+    public Restaurante(String nombre, String apellido, String correo, String contrasenia, LocalDate fechaRegistro, Float calificacion, String nombreRestaurante, Long rut, Direccion direccion, EstadoRestaurante estado, Integer telefono, String descripcion, String cuentaPaypal, String imagen) {
         super(nombre, apellido, correo, contrasenia, fechaRegistro);
         this.calificacion = calificacion;
         this.nombreRestaurante = nombreRestaurante;
